third_party/blink/web_tests/platform/generic/crypto/subtle/rsassa-pkcs1-v1_5/import-jwk-small-key-expected.txt

A key of size smaller than 2048 bits can be used with RS256, RS384, RS512 JWK algorithms.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.algorithm.modulusLength is 1024
PASS key.algorithm.modulusLength is 1024
PASS successfullyParsed is true

TEST COMPLETE
